requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
pydantic==2.12.5
bcrypt==5.0.0
python-multipart==0.0.22
requests==2.32.5
twilio==9.10.3
PyJWT==2.12.1
python-dotenv==1.0.1
tzdata==2026.1